Review and Actor microservices
Split off review and actor functions into their respective microservices and began clean up on movie microservice.
