+++ backend/reel-rating-movie-data-service/src/main/java/edu/oswego/cs/rest/MovieDataService.java
@@ -70,26 +70,6 @@ public class MovieDataService {
     return Response.ok().build();
   }
 
-  @POST
-  @Consumes(MediaType.APPLICATION_JSON)
-  @Path("/actor/create/{movieId}")
-  public Response createActorEndPoint(@Context HttpServletRequest request, Actor actor, @PathParam("movieId") String movieId) throws Exception {
-    String username = getUsername(request);
-    if (username == null) { return Response.status(Response.Status.UNAUTHORIZED).build(); }
-    DatabaseController db = new DatabaseController();
-    db.createActor(actor.getName(), actor.getDateOfBirth(), movieId);
-    return Response.ok().build();
-  }
-  @POST
-  @Consumes(MediaType.APPLICATION_JSON)
-  @Path("/review/create/{movieId}")
-  public Response createReviewEndPoint(@Context HttpServletRequest request, Review review, @PathParam("movieId") String movieId) throws Exception {
-    String username = getUsername(request);
-    if (username == null) { return Response.status(Response.Status.UNAUTHORIZED).build(); }
-    DatabaseController db = new DatabaseController();
-    db.createReview(movieId, review.getReviewDescription(), username, review.getPrivacy());
-    return Response.ok().build();
-  }
 
   /**
    * get endpoints for movies
@@ -197,57 +177,6 @@ public class MovieDataService {
     return Response.ok(movies).build();
   }
 
-  /**
-   * get endpoints for Actors
-   */
-  @GET
-  @Produces(MediaType.APPLICATION_JSON)
-  @Path("/actor/getByName/{name}")
-  public Response getActorByName(@Context HttpServletRequest request, @PathParam("name") String name) throws Exception {
-    String username = getUsername(request);
-    if (username == null) { return Response.status(Response.Status.UNAUTHORIZED).build(); }  
-    DatabaseController dbc = new DatabaseController();
-    List<Actor> actors = dbc.getActorByName(name);
-    return Response.ok(actors).build();
-  }
-
-  /**
-   * get actor by movie ID
-   */
-  @GET
-  @Produces(MediaType.APPLICATION_JSON)
-  @Path("/actor/getByMovieId/{movieId}")
-  public Response getActorByMovieId(@Context HttpServletRequest request, @PathParam("movieId") String movieId) throws Exception {
-    String username = getUsername(request);
-    if (username == null) { return Response.status(Response.Status.UNAUTHORIZED).build(); }
-    DatabaseController dbc = new DatabaseController();
-    List<Actor> actors = dbc.getActorByMovieId(movieId);
-    return Response.ok(actors).build();
-  }
-  /**
-   * get endpoints for reviews
-   */
-  @GET
-  @Produces(MediaType.APPLICATION_JSON)
-  @Path("/reviews/getByUser/{username}")
-  public Response getReviewsByUser(@Context HttpServletRequest request, @PathParam("username") String username) throws Exception {
-    String requesterUsername = getUsername(request);
-    if (requesterUsername == null) { return Response.status(Response.Status.UNAUTHORIZED).build(); }      
-    DatabaseController dbc = new DatabaseController();
-    List<Review> reviews = dbc.getReviewsByUser(username.toLowerCase());
-    return Response.ok(reviews).build();
-  }
-
-  @GET
-  @Produces(MediaType.APPLICATION_JSON)
-  @Path("/reviews/getByMovieId/{movieId}")
-  public Response getReviewsByMovieId(@Context HttpServletRequest request, @PathParam("movieId") String movieId) throws Exception {
-    String requesterUsername = getUsername(request);
-    if (requesterUsername == null) { return Response.status(Response.Status.UNAUTHORIZED).build(); } 
-    DatabaseController dbc = new DatabaseController();
-    List<Review> reviews = dbc.getReviewsByMovieId(movieId);
-    return Response.ok(reviews).build();
-  }
 
   /**
    * Image methods
