+++ backend/reel-rating-review-data-service/src/main/java/edu/oswego/cs/rest/DatabaseController.java
@@ -1,7 +1,9 @@
 package edu.oswego.cs.rest;
 
+import edu.oswego.cs.rest.JsonClasses.Review;
 import org.bson.BsonDateTime;
 import org.bson.Document;
+import org.bson.conversions.Bson;
 import org.bson.types.ObjectId;
 
 import com.mongodb.client.MongoClient;
@@ -9,6 +11,9 @@ import com.mongodb.client.MongoClients;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.Filters;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class DatabaseController {
   String mongoDatabaseName = System.getenv("MONGO_MOVIE_DATABASE_NAME");
@@ -19,6 +24,9 @@ public class DatabaseController {
     return mongoClient.getDatabase(mongoDatabaseName);
   }
 
+  /*
+   * Database get collection methods
+   */
   public MongoCollection<Document> getMovieCollection() {
     return getMovieDatabase().getCollection("movies");
   }
@@ -27,6 +35,19 @@ public class DatabaseController {
     return getMovieDatabase().getCollection("reviews");
   }
 
+  /*
+   * Review Create functions
+   *
+   * createReview
+   */
+  /**
+   * Creates and stores a review in the database. Reviews are the freeform text user generated data. Users are not
+   * allowed to add a review for a movie that does not exist. Users are currently allowed to make multiple reviews
+   * for the same movie.
+   *
+   * @param reviewDescription Freeform text from the user. No limits in size.
+   * @param username the user who created the review
+   */
   public void createReview(String movieIdString, String reviewDescription, String username, String privacy){
     // get collections
     MongoCollection<Document> reviewCollection = getReviewCollection();
@@ -47,5 +68,51 @@ public class DatabaseController {
       reviewCollection.insertOne(newReview);
     }
   }
-  
+
+  /*
+   * Review Get functions
+   *
+   * getReviewsWithFilter
+   *
+   * getReviewsWithMovieId
+   * getReviewsWithUser
+   */
+  private static ArrayList<Review> getReviewsWithFilter(MongoCollection<Document> reviewsCollection, Bson filter) {
+    var reviews = reviewsCollection.find(filter).map(document -> {
+      var re = new Review();
+      re.setUsername(document.getString("username"));
+      re.setReviewDescription(document.getString("reviewDescription"));
+      re.setMovieId(document.getString("movieId"));
+      re.setDateTimeCreated(document.get("dateTimeCreated").toString());
+      re.setPrivacy(document.getString("privacy"));
+      return re;
+    });
+    var list = new ArrayList<Review>();
+    reviews.forEach(list::add);
+    return list;
+  }
+
+  public List<Review> getReviewsWithMovieId(String movieId) {
+    var reviews = getReviewCollection();
+    var filter = Filters.eq("movieId", movieId);
+    return getReviewsWithFilter(reviews, filter);
+  }
+
+  public List<Review> getReviewsWithUser(String username) {
+    var reviews = getReviewCollection();
+    var filter = Filters.eq("username", username);
+    return getReviewsWithFilter(reviews, filter);
+  }
+
+  /*
+   * Review Update functions
+   */
+
+  /*
+   * Review Delete functions
+   */
+
+  /*
+   * Other helper functions
+   */
 }
